Load convention files dynamically
My goal is to be able to add a new convention by adding ONLY a new file
in the conventions/ folder (and a spec file for it of course).

There are 3 parts to this:

1. Move the Convention base module out of the directory.

Although not strictly necessary, this makes it a bit cleaner to simply
treat all of the files in this directory as definitions, not having the
base module sprinkled in.

2. Load the files in the conventions/ directory automatically

This part looks a lot more complicated than it is, I'm fiddling with the
require path to not include the "lib/" part of it, because the load path
is under it, and to not include the ".rb" extension, as this could have
strange side effects - in the end this code requires the exact same
strings as were there explicitly before.

3. Fill the CONVENTION constant with all the required modules
automatically.

This one is self explanatory.